Split `Selector` into its own file so it's easier for users to use and doesn't easily induce cyclic imports.
