+++ tests/unit/test_otel_feedback_computation.py
@@ -2,7 +2,7 @@
 Tests for OTEL Feedback Computation.
 """
 
-from typing import List, Optional
+from typing import List
 
 import pandas as pd
 import pytest
@@ -11,12 +11,12 @@ from trulens.core.otel.instrument import instrument
 from trulens.core.session import TruSession
 from trulens.feedback.computer import MinimalSpanInfo
 from trulens.feedback.computer import RecordGraphNode
-from trulens.feedback.computer import Selector
 from trulens.feedback.computer import _compute_feedback
 from trulens.feedback.computer import _flatten_inputs
 from trulens.feedback.computer import _group_kwargs_by_selectors
 from trulens.feedback.computer import _validate_unflattened_inputs
 from trulens.feedback.computer import compute_feedback_by_span_group
+from trulens.feedback.selector import Selector
 from trulens.otel.semconv.trace import SpanAttributes
 
 from tests.util.mock_otel_feedback_computation import (
@@ -215,137 +215,6 @@ class TestOtelFeedbackComputation(OtelTestCase):
                 eval_root_record_attributes[i][SpanAttributes.EVAL.METRIC_NAME],
                 f"blah{expected_case_number[i]}",
             )
-
-    def test_Selector__split_function_name(self) -> None:
-        self.assertEqual(
-            Selector._split_function_name("a.B.c"), ["a", "B", "c"]
-        )
-        self.assertEqual(
-            Selector._split_function_name("a.py::B::c"), ["a.py", "B", "c"]
-        )
-
-    def test_Selector__matches_function_name(self) -> None:
-        def _test_matches_function_name(
-            selector_function_name: Optional[str],
-            actual_function_name: Optional[str],
-            match_expected: bool,
-        ):
-            selector = Selector(
-                function_name=selector_function_name,
-                span_name="span_name",
-                span_attribute="Z",
-            )
-            if match_expected:
-                self.assertTrue(
-                    selector._matches_function_name(actual_function_name)
-                )
-            else:
-                self.assertFalse(
-                    selector._matches_function_name(actual_function_name)
-                )
-
-        _test_matches_function_name("AA.BB.CC", "AA.BB.CC", True)
-        _test_matches_function_name("BB.CC", "AA.BB.CC", True)
-        _test_matches_function_name("CC", "AA.BB.CC", True)
-        _test_matches_function_name(None, "AA.BB.CC", True)
-        _test_matches_function_name("AAA.BB.CC", "AA.BB.CC", False)
-        _test_matches_function_name("AA.BB.CCC", "AA.BB.CC", False)
-        _test_matches_function_name("A.BB.CC", "AA.BB.CC", False)
-        _test_matches_function_name("B.CC", "AA.BB.CC", False)
-        _test_matches_function_name("C", "AA.BB.CC", False)
-        _test_matches_function_name("AA.BB.CC", None, False)
-        selector = Selector(span_name="AA.BB.CC", span_attribute="Z")
-        self.assertTrue(selector._matches_function_name("AA.BB.CC"))
-        self.assertTrue(selector._matches_function_name("X"))
-        self.assertTrue(selector._matches_function_name(None))
-
-    def test_Selector_matches_spans(self) -> None:
-        selector = Selector(
-            function_name="AA.BB.CC",
-            span_name="XX.YY.ZZ",
-            span_type="span_type",
-            span_attribute="Z",
-        )
-        self.assertTrue(
-            selector.matches_span({
-                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
-                SpanAttributes.SPAN_TYPE: "span_type",
-                "name": "XX.YY.ZZ",
-            })
-        )
-        self.assertFalse(
-            selector.matches_span({
-                SpanAttributes.SPAN_TYPE: "span_type",
-                "name": "XX.YY.ZZ",
-            })
-        )
-        self.assertFalse(
-            selector.matches_span({
-                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
-                "name": "XX.YY.ZZ",
-            })
-        )
-        self.assertFalse(
-            selector.matches_span({
-                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
-                SpanAttributes.SPAN_TYPE: "span_type",
-            })
-        )
-        self.assertTrue(
-            Selector(function_name="CC", span_attribute="Z").matches_span({
-                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
-                SpanAttributes.SPAN_TYPE: "span_type",
-                "name": "XX.YY.ZZ",
-            })
-        )
-        self.assertTrue(
-            Selector(span_name="Y", span_attribute="Z").matches_span({
-                "name": "Y"
-            })
-        )
-
-    def test_Selector_process_span(self) -> None:
-        self.assertEqual(
-            Selector(
-                span_attributes_processor=lambda attributes: "z",
-                function_name="X",
-            ).process_span({}),
-            "z",
-        )
-        self.assertEqual(
-            Selector(span_attribute="Z", function_name="X").process_span({}),
-            None,
-        )
-        self.assertEqual(
-            Selector(span_attribute="Z", function_name="X").process_span({
-                "Z": "z"
-            }),
-            "z",
-        )
-        self.assertEqual(
-            Selector(
-                function_attribute="return", function_name="X"
-            ).process_span({}),
-            None,
-        )
-        self.assertEqual(
-            Selector(
-                function_attribute="return", function_name="X"
-            ).process_span({SpanAttributes.CALL.RETURN: "z"}),
-            "z",
-        )
-        self.assertEqual(
-            Selector(
-                function_attribute="arg1", function_name="X"
-            ).process_span({}),
-            None,
-        )
-        self.assertEqual(
-            Selector(
-                function_attribute="arg1", function_name="X"
-            ).process_span({f"{SpanAttributes.CALL.KWARGS}.arg1": "z"}),
-            "z",
-        )
 
     def test__group_kwargs_by_selectors(self) -> None:
         # Test empty case.
