+++ tests/unit/test_otel_selector.py
@@ -1,0 +1,139 @@
+from typing import Optional
+
+from trulens.core.instruments.otel import SpanAttributes
+from trulens.feedback.selector import Selector
+
+from tests.unit.otel import OtelTestCase
+
+
+class TestOtelSelector(OtelTestCase):
+    def test__split_function_name(self) -> None:
+        self.assertEqual(
+            Selector._split_function_name("a.B.c"), ["a", "B", "c"]
+        )
+        self.assertEqual(
+            Selector._split_function_name("a.py::B::c"), ["a.py", "B", "c"]
+        )
+
+    def test__matches_function_name(self) -> None:
+        def _test_matches_function_name(
+            selector_function_name: Optional[str],
+            actual_function_name: Optional[str],
+            match_expected: bool,
+        ):
+            selector = Selector(
+                function_name=selector_function_name,
+                span_name="span_name",
+                span_attribute="Z",
+            )
+            if match_expected:
+                self.assertTrue(
+                    selector._matches_function_name(actual_function_name)
+                )
+            else:
+                self.assertFalse(
+                    selector._matches_function_name(actual_function_name)
+                )
+
+        _test_matches_function_name("AA.BB.CC", "AA.BB.CC", True)
+        _test_matches_function_name("BB.CC", "AA.BB.CC", True)
+        _test_matches_function_name("CC", "AA.BB.CC", True)
+        _test_matches_function_name(None, "AA.BB.CC", True)
+        _test_matches_function_name("AAA.BB.CC", "AA.BB.CC", False)
+        _test_matches_function_name("AA.BB.CCC", "AA.BB.CC", False)
+        _test_matches_function_name("A.BB.CC", "AA.BB.CC", False)
+        _test_matches_function_name("B.CC", "AA.BB.CC", False)
+        _test_matches_function_name("C", "AA.BB.CC", False)
+        _test_matches_function_name("AA.BB.CC", None, False)
+        selector = Selector(span_name="AA.BB.CC", span_attribute="Z")
+        self.assertTrue(selector._matches_function_name("AA.BB.CC"))
+        self.assertTrue(selector._matches_function_name("X"))
+        self.assertTrue(selector._matches_function_name(None))
+
+    def test_matches_spans(self) -> None:
+        selector = Selector(
+            function_name="AA.BB.CC",
+            span_name="XX.YY.ZZ",
+            span_type="span_type",
+            span_attribute="Z",
+        )
+        self.assertTrue(
+            selector.matches_span({
+                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
+                SpanAttributes.SPAN_TYPE: "span_type",
+                "name": "XX.YY.ZZ",
+            })
+        )
+        self.assertFalse(
+            selector.matches_span({
+                SpanAttributes.SPAN_TYPE: "span_type",
+                "name": "XX.YY.ZZ",
+            })
+        )
+        self.assertFalse(
+            selector.matches_span({
+                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
+                "name": "XX.YY.ZZ",
+            })
+        )
+        self.assertFalse(
+            selector.matches_span({
+                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
+                SpanAttributes.SPAN_TYPE: "span_type",
+            })
+        )
+        self.assertTrue(
+            Selector(function_name="CC", span_attribute="Z").matches_span({
+                SpanAttributes.CALL.FUNCTION: "AA.BB.CC",
+                SpanAttributes.SPAN_TYPE: "span_type",
+                "name": "XX.YY.ZZ",
+            })
+        )
+        self.assertTrue(
+            Selector(span_name="Y", span_attribute="Z").matches_span({
+                "name": "Y"
+            })
+        )
+
+    def test_process_span(self) -> None:
+        self.assertEqual(
+            Selector(
+                span_attributes_processor=lambda attributes: "z",
+                function_name="X",
+            ).process_span({}),
+            "z",
+        )
+        self.assertEqual(
+            Selector(span_attribute="Z", function_name="X").process_span({}),
+            None,
+        )
+        self.assertEqual(
+            Selector(span_attribute="Z", function_name="X").process_span({
+                "Z": "z"
+            }),
+            "z",
+        )
+        self.assertEqual(
+            Selector(
+                function_attribute="return", function_name="X"
+            ).process_span({}),
+            None,
+        )
+        self.assertEqual(
+            Selector(
+                function_attribute="return", function_name="X"
+            ).process_span({SpanAttributes.CALL.RETURN: "z"}),
+            "z",
+        )
+        self.assertEqual(
+            Selector(
+                function_attribute="arg1", function_name="X"
+            ).process_span({}),
+            None,
+        )
+        self.assertEqual(
+            Selector(
+                function_attribute="arg1", function_name="X"
+            ).process_span({f"{SpanAttributes.CALL.KWARGS}.arg1": "z"}),
+            "z",
+        )
